+++ src/math/AlloyTensor3.h
@@ -326,7 +326,7 @@ template<class T, ImageType I> struct Tensor3 {
 				<< im1.dimensions() << "!=" << im2.dimensions());
 		size_t sz = im1.size();
 #pragma omp parallel for
-		for (int offset = 0; offset < (int)sz; offset++) {
+		for (size_t offset = 0; offset < sz; offset++) {
 			func(im1.data[offset], im2.data[offset]);
 		}
 	}
@@ -342,7 +342,7 @@ template<class T, ImageType I> struct Tensor3 {
 				<< im1.dimensions() << "!=" << im2.dimensions());
 		size_t sz = im1.size();
 #pragma omp parallel for
-		for (int offset = 0; offset < (int)sz; offset++) {
+		for (size_t offset = 0; offset < sz; offset++) {
 			func(im1.data[offset], im2.data[offset], im3.data[offset],
 				im4.data[offset]);
 		}
@@ -351,7 +351,7 @@ template<class T, ImageType I> struct Tensor3 {
 		const std::function<void(T&)>& func) {
 		size_t sz = im1.size();
 #pragma omp parallel for
-		for (int offset = 0; offset < (int)sz; offset++) {
+		for (size_t offset = 0; offset < sz; offset++) {
 			func(im1.data[offset]);
 		}
 	}
@@ -364,7 +364,7 @@ template<class T, ImageType I> struct Tensor3 {
 				<< im1.dimensions() << "!=" << im2.dimensions());
 		size_t sz = im1.size();
 #pragma omp parallel for
-		for (int offset = 0; offset < (int)sz; offset++) {
+		for (size_t offset = 0; offset < sz; offset++) {
 			func(im1.data[offset], im2.data[offset]);
 		}
 	}
@@ -377,7 +377,7 @@ template<class T, ImageType I> struct Tensor3 {
 				<< im1.dimensions() << "!=" << im2.dimensions());
 		size_t sz = im1.size();
 #pragma omp parallel for
-		for (int offset = 0; offset < (int)sz; offset++) {
+		for (size_t offset = 0; offset < sz; offset++) {
 			func(im1.data[offset], im2.data[offset], im3.data[offset]);
 		}
 	}
@@ -409,7 +409,7 @@ template<class T, ImageType I> struct Tensor3 {
 				<< im1.dimensions() << "!=" << im2.dimensions());
 		size_t sz = im1.size();
 #pragma omp parallel for
-		for (int offset = 0; offset < (int)sz; offset++) {
+		for (size_t offset = 0; offset < sz; offset++) {
 			func(offset, im1.data[offset], im2.data[offset]);
 		}
 	}
